// applications/applications-web/src/features/lib/application/models/ModelsApplicationCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ModelsApplicationCategory = {
    CategoryPlumber: 'plumber',
    CategoryCarpenter: 'carpenter',
    CategoryElectrician: 'electrician'
} as const;
export type ModelsApplicationCategory = typeof ModelsApplicationCategory[keyof typeof ModelsApplicationCategory];


export function instanceOfModelsApplicationCategory(value: any): boolean {
    for (const key in ModelsApplicationCategory) {
        if (Object.prototype.hasOwnProperty.call(ModelsApplicationCategory, key)) {
            if (ModelsApplicationCategory[key as keyof typeof ModelsApplicationCategory] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ModelsApplicationCategoryFromJSON(json: any): ModelsApplicationCategory {
    return ModelsApplicationCategoryFromJSONTyped(json, false);
}

export function ModelsApplicationCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsApplicationCategory {
    return json as ModelsApplicationCategory;
}

export function ModelsApplicationCategoryToJSON(value?: ModelsApplicationCategory | null): any {
    return value as any;
}

export function ModelsApplicationCategoryToJSONTyped(value: any, ignoreDiscriminator: boolean): ModelsApplicationCategory {
    return value as ModelsApplicationCategory;
}