// applications/applications-web/src/features/lib/application/models/ModelsApplicationEntity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ModelsApplicationCategory } from './ModelsApplicationCategory';
import {
    ModelsApplicationCategoryFromJSON,
    ModelsApplicationCategoryFromJSONTyped,
    ModelsApplicationCategoryToJSON,
    ModelsApplicationCategoryToJSONTyped,
} from './ModelsApplicationCategory';

/**
 * 
 * @export
 * @interface ModelsApplicationEntity
 */
export interface ModelsApplicationEntity {
    /**
     * 
     * @type {ModelsApplicationCategory}
     * @memberof ModelsApplicationEntity
     */
    category: ModelsApplicationCategory;
    /**
     * 
     * @type {string}
     * @memberof ModelsApplicationEntity
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsApplicationEntity
     */
    description: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsApplicationEntity
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ModelsApplicationEntity
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsApplicationEntity
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsApplicationEntity
     */
    updatedAt?: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsApplicationEntity
     */
    userId?: number;
}



/**
 * Check if a given object implements the ModelsApplicationEntity interface.
 */
export function instanceOfModelsApplicationEntity(value: object): value is ModelsApplicationEntity {
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function ModelsApplicationEntityFromJSON(json: any): ModelsApplicationEntity {
    return ModelsApplicationEntityFromJSONTyped(json, false);
}

export function ModelsApplicationEntityFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsApplicationEntity {
    if (json == null) {
        return json;
    }
    return {
        
        'category': ModelsApplicationCategoryFromJSON(json['category']),
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'description': json['description'],
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'],
        'status': json['status'] == null ? undefined : json['status'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
        'userId': json['user_id'] == null ? undefined : json['user_id'],
    };
}

export function ModelsApplicationEntityToJSON(json: any): ModelsApplicationEntity {
    return ModelsApplicationEntityToJSONTyped(json, false);
}

export function ModelsApplicationEntityToJSONTyped(value?: ModelsApplicationEntity | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'category': ModelsApplicationCategoryToJSON(value['category']),
        'created_at': value['createdAt'],
        'description': value['description'],
        'id': value['id'],
        'name': value['name'],
        'status': value['status'],
        'updated_at': value['updatedAt'],
        'user_id': value['userId'],
    };
}