// applications/applications-web/src/features/lib/application/models/ServiceApplicationCreateDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document:
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { ModelsApplicationCategory } from "./ModelsApplicationCategory";
import {
  ModelsApplicationCategoryFromJSON,
  ModelsApplicationCategoryFromJSONTyped,
  ModelsApplicationCategoryToJSON,
  ModelsApplicationCategoryToJSONTyped,
} from "./ModelsApplicationCategory";

/**
 *
 * @export
 * @interface ServiceApplicationCreateDto
 */
export interface ServiceApplicationCreateDto {
  /**
   *
   * @type {ModelsApplicationCategory}
   * @memberof ServiceApplicationCreateDto
   */
  category: ModelsApplicationCategory;
  /**
   *
   * @type {string}
   * @memberof ServiceApplicationCreateDto
   */
  description: string;
  /**
   *
   * @type {string}
   * @memberof ServiceApplicationCreateDto
   */
  name: string;
}

/**
 * Check if a given object implements the ServiceApplicationCreateDto interface.
 */
export function instanceOfServiceApplicationCreateDto(
  value: object,
): value is ServiceApplicationCreateDto {
  if (!("category" in value) || value["category"] === undefined) return false;
  if (!("description" in value) || value["description"] === undefined)
    return false;
  if (!("name" in value) || value["name"] === undefined) return false;
  return true;
}

export function ServiceApplicationCreateDtoFromJSON(
  json: any,
): ServiceApplicationCreateDto {
  return ServiceApplicationCreateDtoFromJSONTyped(json, false);
}

export function ServiceApplicationCreateDtoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ServiceApplicationCreateDto {
  if (json == null) {
    return json;
  }
  return {
    category: ModelsApplicationCategoryFromJSON(json["category"]),
    description: json["description"],
    name: json["name"],
  };
}

export function ServiceApplicationCreateDtoToJSON(
  json: any,
): ServiceApplicationCreateDto {
  return ServiceApplicationCreateDtoToJSONTyped(json, false);
}

export function ServiceApplicationCreateDtoToJSONTyped(
  value?: ServiceApplicationCreateDto | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    category: ModelsApplicationCategoryToJSON(value["category"]),
    description: value["description"],
    name: value["name"],
  };
}